Replace size_t type alias with libc::size_t
This is necessary for building on 32bit systems.

+++ fuzz/fuzz_targets/decompress.rs
@@ -12,8 +12,7 @@ fn rs(compressed: &[u8]) -> (usize, Vec<u8>) {
     let compressed_size = compressed.len();
 
     // Get decompressed size from frame header
-    let decompressed_size =
-        unsafe { ZSTD_getFrameContentSize(compressed_ptr, compressed_size as u64) };
+    let decompressed_size = unsafe { ZSTD_getFrameContentSize(compressed_ptr, compressed_size) };
     if decompressed_size == ZSTD_CONTENTSIZE_ERROR {
         return (decompressed_size as usize, vec![]);
     } else if decompressed_size == ZSTD_CONTENTSIZE_UNKNOWN {
@@ -25,9 +24,9 @@ fn rs(compressed: &[u8]) -> (usize, Vec<u8>) {
     let result = unsafe {
         ZSTD_decompress(
             decompressed.as_mut_ptr() as *mut c_void,
-            decompressed.len() as u64,
+            decompressed.len(),
             compressed_ptr,
-            compressed_size as u64,
+            compressed_size,
         )
     };
